requirements.txt
streamlit
pandas
pypdf
altair
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
requests
